requirements.txt
pandas
pandas-gbq
google-cloud-bigquery
kaggle
pyarrow
jupyter
nbconvert
ipykernel
